requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
flask-bcrypt